Always load defkeymap
It seems that there's a bug in the kernel 2.6.24,
which makes the built-in defkeymap to produce incorrect composite chars.
Thus keymaps that don't have 'include "compose.latin1"' have issues (FS#9593).
While it is better to fix this bug in kernel and all affected keymaps
loading defkeymaps before other keymaps seems to not hurt anyway.

Signed-off-by: Roman Kyrylych <roman@archlinux.org>