Shutdown non-rc.conf daemons first at rc.single
To match with commit 7f1bd75ee6aaffddf47e704c46fbab5396b7dd19
for rc.shutdown.

Flip the order in which daemons are shutdown. First we scan for
all daemons that are NOT in rc.conf and shut those down (order is
based on ls, so it is dependent on LC_COLLATE). Afterwards, we
shut down all daemons in rc.conf, in order.

Signed-off-by: Gerardo Exequiel Pozzi <vmlinuz386@yahoo.com.ar>
Signed-off-by: Aaron Griffin <aaronmgriffin@gmail.com>