Fix stty/tput usage for columns again
Cleans up double application of earlier fixes for this issue.

Instead of calling $(tput cols 2>/dev/null), we now try tput cols
silently to see whether it complains. If not, then we call $(tput cols)
with no redirection of stderr. This way, we can get results other than
80.

eliminate color when stdout is not terminal

Signed-off-by: Jim Pryor <profjim@jimpryor.net>
Signed-off-by: Aaron Griffin <aaronmgriffin@gmail.com>