Bashify initscripts.
We rely on bash specific features (arrays and associative arrays).
Trying to maintain POSIX compatibility is a net performancle loss, since native bash
constructs tend to perform better and lend themselves to less buggy code.

Start off by adding #!/bin/bash to the top of functions to make it clear that we
are not POSIX, and to help text editors perform appropriate syntax highlighting.

Tighten up the console size finding code a bit, and simplify the code that clears
USECOLOR.

Use [[ ]] instead of [ ] for conditional checking when running in bash.
It is worth 10 - 30% speedup whenever you want to compare something.

Instead of calling a command and then testing for nonzero exit status, just
test the command exit status directly.