Rewrite early RTC device creation
Creating the RTC device will result in an error if devtmpfs is used,
as it already exists after loading the module. The new code skips
the mknod if the device is already present.

Additionally to rtc-cmos, we now also try to load rtc and genrtc,
as some custom kernels use the "old" misc RTC device instead of
the newer RTC class. In this case, the created device is also
different. This should fix FS#18078.