cryptsetup: read keys for non-root LUKS partitions as raw data from block device
For LUKS encrypted root partitions it is possible to store a keyfile hidden as raw data onto a USB stick with the dd command. This has the advantage, that the key is saved on a USB stick, but can't be found be just mounting the USB stick.

The appropriate line in /etc/crypttab is similar to the kernel parameter in /boot/grub/menu.lst:
home/dev/sdaXBLOCKDEVICE:OFFSET:SIZE

[teg: fixed whitespace and expanded paths to binaries]
Signed-off-by: Tom Gundersen <teg@jklm.no>