PATH: export a standard path in rc.sysinit
This is usefull for udev and other processes that do not get a
PATH set by /etc/profile.

The particular choice of paths to export was taken from systemd,
the rationale being that systemd is designed to be compatible with
all the major distros and in particular with udev.

The paths are also the same as set as standard in /etc/profile

With this patch we handle the PATH variable similarly to what systemd
does (in their case the variable is set in init).

Signed-off-by: Tom Gundersen <teg@jklm.no>