Use modprobe's --all flag instead of looping.
Filter blacklisted modules with a parameter expansion and add an extra
conditional to modprobe to only act if the resultant expansion has a
value. It's important that no quoting takes place here, or else a
variable is defined with a null terminator and which will cause the
extra conditional to pass and modprobe will error out.

Signed-off-by: Dave Reisner <d@falconindy.com>
Signed-off-by: Tom Gundersen <teg@jklm.no>