dev: create pts and shm directories
Creates the mountpoints for the standard api filesystems in /etc/fstab.

At the moment this is already done by /lib/udev/devices in our udev
package. However, with this change it is possible to boot a fully
working system with the udev unaltered from upstream.

For the future, this patch allows us to get rid of the /lib/udev/devices
directory from the udev package alltogether.

v2: simplified after suggestion from Dave.

Signed-off-by: Tom Gundersen <teg@jklm.no>