lvm/raid/crypt/btrfs: remove support
The aim of this package is to supply a minimalistic boot. The old
handling of lvm/raid/crypt/btrfs is too brittle and it is out of scope
of this project.

Note that btrfs can still be used, just not over multiple devices. There
are plans to fix this upstream.