mtab: don't write to mtab if it is a symlink
In the future /etc/mtab should be a symlink to /proc/self/mounts, add
support for this in rc.sysinit to be prepared. While strictly not
needed, this makes it clear that the functionality can be removed once
we move over to using symlinks.

This functionality is expected with the next release of util-linux
(2.20), but is already available in the current version if the
experimental libmount support is enabled.

v2: simplified after suggestion from Seblu.

Signed-off-by: Tom Gundersen <teg@jklm.no>