hwclock: do not allow RTC to be in localtime
Using RTC in localtime will cause problems. For two reasons:
1) general: due to DST, localtime is a flawed concept; it is not well
defined. I.e. there are several points in time correpsonding to one
localtime (for one hour every year). This can cause problems that
are impossible to solve, even if they will occur relatively infrequently.
2) linux specific: when the RTC device is initialized by the kernel the
system time is set automatically. The kernel assumes that the RTC device
is in UTC, and there is no way to override this. If the RTC is actually
in localtime, this will mean there is a window of time between the rtc
device is initialized and hwclock is called where the system time is
wrong (by up to 12 hours), this might cause intermittent and hard to
debug problems.

The second point could be overcome by patching the kernel to
2.1) no longer set the system time automatically, I assume this is not
going to be accepted, as there must have been a reason they made the
change initially (and as pointed out above using anything else than UTC
is crazy).
2.2) take a boot parameter that indicates the timezone. In this case, we
should read this parameter directly, rather than reading rc.conf, as
having rc.conf and the parameter differ would cause havoc.

Lastly, the only reason to use localtime is in a dualboot scenario with
Windows (all other operating systems, to the best of my knowledge, use
UTC). However, one can configure Windows to use UTC (solving the kinds
of problems described above), so this is not really a valid concern.

tl;dr: allowing RTC to be in localtime is fundamentally broken, and
hence not the Arch way ;-)

Signed-off-by: Tom Gundersen <teg@jklm.no>