timezone: do not set timezone from rc.conf
The timezone in rc.conf is not necessarily in sync with /etc/timezone,
which might be changed by other tools (e.g., GUIs) during runtime. We
cannot know whether or not /etc/timezone or /etc/rc.conf is the correct
one.

There is no need to set this in rc.conf when the user could just copy
the timezone file manually, or use a GUI tool.

Signed-off-by: Tom Gundersen <teg@jklm.no>