/dev: mount /dev/pts and /dev/shm during early boot
This has two purposes:

1) it allows boot services like bootlogd (that needs /dev/pts) to use
these folders.
2) by setting the standard options in rc.sysinit rather than in fstab we
have control over them and can easily integrate them with other packages
and make adjustments to fix bugs.

I could not think of any usecases where someone might want to override
the default settings, but if there is a need we could add support for
this too (I think we should avoid this as much as possible, but if there
are valid reasons I will not object).

In the future we should probably also put /run{,/lock} and (once they
can safely be put on tmpfs) /var/{run,lock} under the control of
rc.sysinit.

Signed-off-by: Tom Gundersen <teg@jklm.no>