/run: mount /run if it is not already mounted
This is in accordance with the new "standard" agreed on by the major
distros, and discussed on our mailinglists.

This folder will be used by udev/mdadm and other tools rather than
putting things in /dev/.XXX. In the future we might symlink
/var/{run,lock} to /run/{,lock} as the other distros are intending to
do. First, we would have to make sure that /var/{run,lock} can be
mounted as tmpfs without problems.

Signed-off-by: Tom Gundersen <teg@jklm.no>