bootlogd: log messages to console during early boot
This saves all the output during boot to /var/log/boot, which makes
debugging boot problems much simpler.

The logfile contains all the control (color) characters from boot, so
some post processing is in order before the log can be read in a nice
format (e.g. less should be able to show this nicely in raw mode).

Start bootlogd as soon as the required /dev/pts is mounted.
When boot is finished (either at the end of rc.multi or rc.single)
create the required lof file (/var/log/boot) if it does not exist. This
causes bootlogd to flush the log to disk.
Finally, send SIGTERM to bootlogd and clean up after it.

Idea-and-original-patch-by: Søren Poulsen <nikorpoulsen@gmail.com>
Tested-by: David Goldstein <djg1971@gmail.com>
Signed-off-by: Tom Gundersen <teg@jklm.no>