create /run/lock
This allows /var/run and /var/lock to be symlinked to /run and
/run/lock, respectively. This would be the final step in fixing
FS#18157.

With /run/lock created early in boot it would be possible for tools
like mdadm and lvm2 to use this directory for their locking, and
maybe avoid the --sysinit parameter. This would have to be implemented
in the respective packages though.

Finally, /run/lock is created 1777 just like /var/lock is now.
For the future, it might make sense to follow systemd's lead and make
it 0755, and then create a subdir /run/lock/lockdev owned by root:lock
with permissions 0755. This would improve security and fend agains DoS
attacks, but we would have to check carefully what packages are affected.

Signed-off-by: Tom Gundersen <teg@jklm.no>