rc: exit with error count, not error sum
cc199761f assumes that /etc/rc.d scripts will exit with a value of 0 or
1. Since this can't be guaranteed, clamp the return value to 0 or 1
before adding it to the exit value.

Note that a lot of /etc/rc.d scripts don't properly exit with 0 or 1, so
this is more of a forward looking change in the hope that these scripts
are cleaned up.

Signed-off-by: Tom Gundersen <teg@jklm.no>