Improve coloring on light and transparent term
This patch improve coloring in light and transparent term.
It uses tput if available or default color code.

It remplace \033 by \e. It's a kind of rollback from commit 65f410,
because initscripts are now in bash, and argument
"\e isn't recongized in all shells" is not an issue anymore.

This is not done in locale.sh script which can be sourced by others shell,
so we try to be the more compliant as possible

Signed-off-by: Sebastien Luttringer <seblu@seblu.net>