releasing: update instructinos after going through it once
I struggled a bit with finding the simplest way of creating packages, so after asking on irc, I thought I'd add the answer here.

This is not really initscripts specific, so could easiy be expanded and put in a wiki. Should work for any package that can be built without also building its deps.

Signed-off-by: Tom Gundersen <teg@jklm.no>