Revert "udev: generate blacklist on boot"
This reverts commit 99d64f5789f01243903be4b32078ec38bdce75fa.

We decided not to provide a compatibility layer, but rather require
the user to create blacklist files in /etc/modprobe.d manually.
Hence, this is no longer needed.

The reason was that the compatibliy would not be 1-to-1, so it might
be missleading to pretend that it was. Also, most people should not
need the blacklisting in the first place.

Conflicts:

	Makefile

Signed-off-by: Tom Gundersen <teg@jklm.no>