hwclock: allow TZ to be set in adjtime, only adjust for TZ during boot
This allows HARDWARECLOCK to be removed from rc.conf, if you want to
rely on /var/lib/hwclock/adjtime (or wherever the file ends up in the
future). It assumes the relevant path is on the rootfs.

Furthermore, we can assume (since kernel2.6.??) that the rtc modules
set the time from rtc themselves, so our job is just to adjust the tz.
(FWIW, this coincides with what systemd does).

To get the old behavior of not touching the rtc, set it to anything
anything else.

We should probably have chosen a syntax for this, like IGNORE. Maybe
another time...

Signed-off-by: Tom Gundersen <teg@jklm.no>