rc.sysinit: Move 'run_hook sysinit_start' after mount root read-only
Rationale

Currently we have to duplicate at least the mount /proc code line into a
custom sysinit_start hook to be able to start a splash daemon or something
in case no initcpio is used.
Moreover we want to make sure nothing can open files read/write on root
which would block the read-only remount.
So running the sysinit_start hook a bit later appears to be more save.