udev: exit on shutdown
udev was receiving some events (due to e.g. swapoff) on shutdown
that caused it to fork new processes. These then receivde TERM before
they could finish, and complained on the console.

In principle, I'm worried that, with the right ammonut of bad luck,
we could fork off some process at exactly the wrong time which escapes
the killall logic.

This, by the way, highlights the frailty of the killall stuff.

Signed-off-by: Tom Gundersen <teg@jklm.no>