rc.shutdown: reorder a bit
The new ordering is now:

1) book-keeping
2) shutting down and killing processes
3) swapoff
4) unmounting and tearing down of block devices

The reason to move 1) to the top is so it will not interfere with
the killing of processes. The reason for not moving 3) to the top
as well is that we want to minimize the amount of stuff that has to
be moved from swap to ram.

However, swapoff must happen before 4) as we don't know what is under
the swap device/file.

Added a TODO: unmount any devices backed by swap (tmpfs/ramfs) to make
sure that their contents is thrown away rather than moved to ram. In
principle we might get OOM on shutdown due to the current situation,
but in most cases the only problem should be that shutdown is unecessarily
slow.

Thanks to Dan and Dave for prompting me to look into this.

Cc: Dan McGee <dpmcgee@gmail.com>
Cc: Dave Reisner <d@falconindy.com>
Signed-off-by: Tom Gundersen <teg@jklm.no>